python_server/requirements.txt
asgiref==3.4.1
Django==3.2.6
django-cors-headers==3.7.0
lxml==4.6.3
Pillow==8.3.1
python-pptx==0.6.19
pytz==2021.1
sqlparse==0.4.1
XlsxWriter==1.4.5